+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/transaction/OutboundTransactionManagerImpl.kt
@@ -4,7 +4,9 @@ import cash.z.ecc.android.sdk.internal.Twig
 import cash.z.ecc.android.sdk.internal.ext.toHexReversed
 import cash.z.ecc.android.sdk.internal.model.EncodedTransaction
 import cash.z.ecc.android.sdk.model.Account
+import cash.z.ecc.android.sdk.model.Proposal
 import cash.z.ecc.android.sdk.model.TransactionRecipient
+import cash.z.ecc.android.sdk.model.TransactionSubmitResult
 import cash.z.ecc.android.sdk.model.UnifiedSpendingKey
 import cash.z.ecc.android.sdk.model.Zatoshi
 import co.electriccoin.lightwallet.client.LightWalletClient
@@ -41,18 +43,40 @@ internal class OutboundTransactionManagerImpl(
         }
     }
 
-    override suspend fun submit(encodedTransaction: EncodedTransaction): Boolean {
+    override suspend fun proposeTransfer(
+        account: Account,
+        recipient: String,
+        amount: Zatoshi,
+        memo: String
+    ): Proposal = encoder.proposeTransfer(account, recipient, amount, memo.toByteArray())
+
+    override suspend fun proposeShielding(
+        account: Account,
+        memo: String
+    ): Proposal = encoder.proposeShielding(account, memo.toByteArray())
+
+    override suspend fun createProposedTransactions(
+        proposal: Proposal,
+        usk: UnifiedSpendingKey
+    ): List<EncodedTransaction> = encoder.createProposedTransactions(proposal, usk)
+
+    override suspend fun submit(encodedTransaction: EncodedTransaction): TransactionSubmitResult {
         return when (val response = service.submitTransaction(encodedTransaction.raw.byteArray)) {
             is Response.Success -> {
                 if (response.result.code == 0) {
                     Twig.debug { "SUCCESS: submit transaction completed" }
-                    true
+                    TransactionSubmitResult.Success(encodedTransaction.txId)
                 } else {
                     Twig.debug {
                         "FAILURE! submit transaction ${encodedTransaction.txId.byteArray.toHexReversed()} " +
                             "completed with response: ${response.result.code}: ${response.result.message}"
                     }
-                    false
+                    TransactionSubmitResult.Failure(
+                        encodedTransaction.txId,
+                        false,
+                        response.result.code,
+                        response.result.message
+                    )
                 }
             }
 
@@ -62,7 +86,12 @@ internal class OutboundTransactionManagerImpl(
                         response.description
                     }"
                 }
-                false
+                TransactionSubmitResult.Failure(
+                    encodedTransaction.txId,
+                    true,
+                    response.code,
+                    response.description
+                )
             }
         }
     }